Migrated the other projects to the new test structure. Updated NuGet package descriptions.
