--- src/NCommons.Collections.Tests/PriorityQueue/ConstructorTests.cs
@@ -1,19 +1,0 @@
-﻿namespace NCommons.Collections.Tests.PriorityQueue
-{
-    using System.Collections.Generic;
-    using Xunit;
-
-    public class ConstructorTests
-    {
-
-        [Fact]
-        public void Can_Pass_Null()
-        {
-            // Should not throw.
-            new PriorityQueue<int>((IComparer<int>)null);
-            new PriorityQueue<int>((IPriorityQueue<int>)null);
-        }
-
-    }
-
-}
